Fix command path
