+++ tools/dfx-orbit/src/args/request/canister/call.rs
@@ -27,7 +27,7 @@ pub struct RequestCanisterCallArgs {
 }
 
 impl RequestCanisterCallArgs {
-    /// Converts the CLI arg type into the equivalent Orbit API type.
+    /// Converts the CLI arg stype into the equivalent Orbit API type.
     pub(crate) fn into_create_request_input(
         self,
         station_agent: &StationAgent,
@@ -39,6 +39,7 @@ impl RequestCanisterCallArgs {
             argument,
         } = self;
         let canister_id = station_agent.canister_id(&canister)?;
+
         // TODO: It would be really nice to be able to use `blob_from_arguments(..)` here, as in dfx, to geta ll the nice things such as help composing the argument.
         let arg = if let Some(argument) = argument {
             Some(
